+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow playing paused torrents if file is completed.
 
+### Fixed
+
+- Properly handle invalid range offsets.
+
 ## [v0.0.8] - 07/05/2025
 
 ### Changed